+++ Tests/Book.swift
@@ -17,7 +17,7 @@ final class Book : Transactable, CustomStringConvertible {
 
     private var _transactionContext: TransactionContext? = nil
 
-    private var _isCommittable = true
+    private var _committabilityError: Error? = nil
 
     var transactionContext: TransactionContext {
         return _transactionContext!
@@ -38,8 +38,8 @@ final class Book : Transactable, CustomStringConvertible {
         _transactionContext = TransactionContext.createNode(owner: self, parent: library)
     }
 
-    func isCommittable() -> Bool {
-        return _isCommittable
+    func hasCommittabilityError() -> Error? {
+        return _committabilityError
     }
 
     func onBegin(transaction: Transaction) {
@@ -60,8 +60,8 @@ final class Book : Transactable, CustomStringConvertible {
         rollbackCount = 0
     }
 
-    func setCommittable(_ isCommittable: Bool) {
-        _isCommittable = isCommittable
+    func setCommittabilityError(_ error: Error) {
+        _committabilityError = error
     }
 
 }