+++ Tests/Test_ActiveTransaction.swift
@@ -80,9 +80,9 @@ class Test_ActiveTransaction: XCTestCase {
     }
 
     final func test_InitialState_TransactablesAreCommittable() {
-        XCTAssertTrue(library.isCommittable())
-        XCTAssertTrue(bookA.isCommittable())
-        XCTAssertTrue(bookB.isCommittable())
+        XCTAssertNil(library.hasCommittabilityError())
+        XCTAssertNil(bookA.hasCommittabilityError())
+        XCTAssertNil(bookB.hasCommittabilityError())
     }
 
 }