+++ Tests/Test_ActiveTransaction_NodeContext.swift
@@ -18,7 +18,7 @@ class Test_ActiveTransaction_NodeContext : Test_ActiveTransaction {
             guard let error = $0 as? TransactionError else {
                 XCTFail("Must throw TransactionError, but thrown \(type(of: $0)) instead"); return
             }
-            XCTAssertEqual(error, TransactionError.anotherTransactionIsStillActive)
+            XCTAssertEqual(error, TransactionError.anotherTransactionIsActive)
         }
     }
 
@@ -33,18 +33,18 @@ class Test_ActiveTransaction_NodeContext : Test_ActiveTransaction {
     }
 
     func test_CommitTransaction_WithWildcardDescriptor_OnNoncommittable_Throws() {
-        library.setCommittable(false)
+        library.setCommittabilityError(LibraryError.someError)
 
         XCTAssertThrowsError(try self.bookA.commitTransaction()) {
-            guard let error = $0 as? TransactionError else {
-                XCTFail("Must throw TransactionError, but thrown \(type(of: $0)) instead"); return
+            guard let error = $0 as? LibraryError else {
+                XCTFail("Must throw LibraryError, but thrown \(type(of: $0)) instead"); return
             }
-            XCTAssertEqual(error, TransactionError.uncommittableTransaction)
+            XCTAssertEqual(error, LibraryError.someError)
         }
     }
 
     func test_CommitTransaction_WithWildcardDescriptor_OnNoncommittable_DoesNotTriggerTransactable() {
-        library.setCommittable(false)
+        library.setCommittabilityError(LibraryError.someError)
 
         do { try bookB.commitTransaction(noncurrentTransaction) } catch { }
 
@@ -70,18 +70,18 @@ class Test_ActiveTransaction_NodeContext : Test_ActiveTransaction {
     }
 
     func test_CommitTransaction_WithCurrentDescriptor_OnNoncommittable_Throws() {
-        library.setCommittable(false)
+        library.setCommittabilityError(LibraryError.someError)
 
         XCTAssertThrowsError(try self.bookB.commitTransaction(currentTransaction)) {
-            guard let error = $0 as? TransactionError else {
-                XCTFail("Must throw TransactionError, but thrown \(type(of: $0)) instead"); return
+            guard let error = $0 as? LibraryError else {
+                XCTFail("Must throw LibraryError, but thrown \(type(of: $0)) instead"); return
             }
-            XCTAssertEqual(error, TransactionError.uncommittableTransaction)
+            XCTAssertEqual(error, LibraryError.someError)
         }
     }
 
     func test_CommitTransaction_WithCurrentDescriptor_OnNoncommittable_DoesNotTriggerTransactable() {
-        library.setCommittable(false)
+        library.setCommittabilityError(LibraryError.someError)
 
         do { try bookA.commitTransaction(currentTransaction) } catch { }